+++ R/tidy_genomic_data.R
@@ -34,7 +34,8 @@
 #' @param whitelist.markers (optional) A whitelist containing CHROM (character
 #' or integer) and/or LOCUS (integer) and/or
 #' POS (integer) columns header. To filter by chromosome and/or locus and/or by snp.
-#' The whitelist is in the working directory (e.g. "whitelist.txt").
+#' The whitelist is an object in your
+#' global environment or a file in the working directory (e.g. "whitelist.txt").
 #' de novo CHROM column with 'un' need to be changed to 1.
 #' In the VCF, the column ID is the LOCUS identification.
 #' Default \code{whitelist.markers = NULL} for no whitelist of markers.
@@ -51,8 +52,10 @@
 #' a combination of 2 (e.g. CHROM and POS or CHROM and LOCUS or LOCUS and POS) or
 #' all 3 (CHROM, LOCUS, POS). The markers columns must be designated: CHROM (character
 #' or integer) and/or LOCUS (integer) and/or POS (integer). The id column designated
-#' INDIVIDUALS (character) columns header. The blacklist must be in the working
-#' directory (e.g. "blacklist.genotype.txt"). For de novo VCF, CHROM column
+#' INDIVIDUALS (character) columns header.
+#' The blacklist is an object in your global environment or
+#' a file in the working directory (e.g. "blacklist.genotype.txt").
+#' For de novo VCF, CHROM column
 #' with 'un' need to be changed to 1.
 #' Default: \code{blacklist.genotype = NULL} for no blacklist of
 #' genotypes to erase.
@@ -1528,8 +1531,11 @@ The POS column used in the MARKERS column is different in biallelic and multiall
   } else {
     if (verbose) message("Erasing genotype: yes")
     want <- c("MARKERS", "CHROM", "LOCUS", "POS", "INDIVIDUALS")
+    if (is.vector(blacklist.id)) {
     suppressWarnings(suppressMessages(
-      blacklist.genotype <- readr::read_tsv(blacklist.genotype, col_names = TRUE) %>%
+      blacklist.genotype <- readr::read_tsv(blacklist.genotype, col_names = TRUE)))
+    }
+    suppressWarnings(suppressMessages(blacklist.id <- blacklist.id %>%
         dplyr::mutate(
           INDIVIDUALS = stringi::stri_replace_all_fixed(
             str = INDIVIDUALS,